Run elm-format and 'mix format' from project root (#36)
Some formatters need to be run from the project root directory in
order to find their settings. Support this by adding (yet another)
argument, root-files, to the format-all-buffer-hard function.

We need to be able to support multiple alternative filenames and stop
looking in parent directories as soon as any one of them is found.
elm-format relies on this.

elm-format is being fixed upstream to look for the right files itself
so it doesn't need this support from us. Fix status for 'mix format'
is not known; mix is a 'make'-style tool so I don't know whether they
want to support running from a directory other than the project root.

Fixes #36.