Fix scalafmt output issue
The homebrew version of scalafmt is out of date and not the recommended way to
install scalafmt. It does not properly support the `--quiet` flag, leading to
output of `Reformatting...` on `stderr`, which gets reported as an error. The
most recent version fixes this, and can be installed cross platform using
`coursier`.