scalafmt: Add --stdout as workaround
Sometimes, scalafmt does not produce formatted output to stdout with
only the `--stdin` flag provided. For example, this happens when there
are no formatting changes required to be made to the file. As-is, this
causes the formatter to clear the buffer when there are no formatting
changes required.

As a workaround, adding the `--stdout` flag ensure that output is always
put on stdout.

See https://github.com/scalameta/scalafmt/issues/3632 for more
information.